requirements_testing.txt
# Usage: pip install -r requirements.txt
coverage
# Required to run browser tests:
selenium
pyvirtualdisplay
django-debug-toolbar >= 3.2.2, < 4
prospector
pytest-playwright
